+++ CHANGES.md
@@ -27,6 +27,7 @@
 - Added additional voucher restrictions
 - Added groups claim for OpenID
 - Minecraft player info now shows previous names
+- Show warnings on the STEMCraft punishments page
 
 ## 1.4.3